% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.fload_sfc}
\alias{.fload_sfc}
\title{Load GeoJSON Files}
\usage{
.fload_sfc(x, expand_geometries = FALSE)
}
\arguments{
\item{x}{GeoJSON file}

\item{expand_geometries}{default: \code{FALSE}}
}
\description{
Load GeoJSON Files
}
