% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.fparse_sfc}
\alias{.fparse_sfc}
\title{Parse GeoJSON Strings}
\usage{
.fparse_sfc(x, expand_geometries = FALSE)
}
\arguments{
\item{x}{GeoJSON string}

\item{expand_geometries}{default: \code{FALSE}}
}
\description{
Parse GeoJSON Strings
}
